requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
charset-normalizer==3.3.0
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
psutil==5.9.6
python-dotenv==1.0.0
yarl==1.9.2
